+++ CHANGELOG.md
@@ -30,6 +30,8 @@ is available in the [commit logs](https://github.com/scverse/scvi-tools/commits/
 -   Add `unsigned` argument to {meth}`scvi.hub.HubModel.pull_from_s3` to allow for unsigned
     downloads of models from AWS S3 {pr}`2615`.
 -   Add support for `batch_key` in {meth}`scvi.model.CondSCVI.setup_anndata` {pr}`2626`.
+-   Add `load_best_on_end` argument to {class}`scvi.train.SaveCheckpoint` to load the best model
+    state at the end of training {pr}`xxxx`.
 
 #### Changed